rename solver ast to clause
