+++ antwort-solver/src/lib.rs
@@ -1,2 +1,2 @@
-pub mod ast;
+pub mod clause;
 pub mod error;
